feat: Add login
This connects the app to a firebase project and adds logging in. On
startup, the app will now decide whether we need to log in again and
prompt the user for their credentials if needed. Otherwise it will go
to a dummy home screen. After when logging in, the app validates the
credentials with the help of firebase authentication and then brings
the user to the home screen as well.