feat: Home screen layout and subjects widget
The home screen will now choose a layout depending on whether we are in
landscape or portrait mode and lay out two children accordingly:
- landscape: horizontal layout (next to each other)
- portrait: vertical layout (on top of each other)

The widgets laid out are a soon-due tasks widget and a subjects widget.
The former currently only has a placeholder, the latter has the UI
defined and filled with some example data.