feat: Add page to create tasks
This adds a new page, available via the plus in the bottom left corner
on the home screen, allowing you to create tasks. It requires a title,
a due date, a reminder and a subject. Optionally you can provide a
description. On confirmation, the task is created in the firestore
database.