feat: Add SubjectPage
This adds a new page which is accessible through clicking on the
subject in the grid on the home screen. In this screen, you can see all
tasks associated with this subject, as well as edit + delete the
subject. In the future, this screen's layout will look similar to the
one of the home screen, and thus show the available documents for the
subject.