refactor: Inject database implementation using GetIt
This commit introduces a database interface, defining the functions
needed for the app to work. The original database class has been
renamed to DatabaseFirestore. It implements the Database interface and
thus provides an interface to communicate with Firestore. In Setup,
this class is registered using the "get_it" package, and the registered
implementation can be retrieved using Database.I (which in turn asks
get_it for the registered class).
In the future, this can be used in inject a different Database
implementation to make using the app without an account possible.