feat: Add option of not using an account
On first login, the user now has the option whether they want to create
an account or continue without one for now. When continuing without
one, the app remembers that state and does not prompt the login page
when starting the app. Additionally, the app uses a different
implementation of the Database class in that case, which refers to a
local sqlite database.
When the user does not have an account, they are given the option to
register at the top right of the home screen (where the logout button
would be). Once the user creates an account, any locally stored data
from the sqlite database is automatically pushed to Firestore.