feat(AccountDialog): Add feature to delete the account
In the AccountDialog, there is now a button with which the user can
delete their account. This requires the user to type in their password
again to reauthenticate, after which it will delete all the data of the
user (subjects + tasks), delete the account and finally sends them back
to the login screen.
From here on, the user may also use the app without an account.