feat(Subject): Add task count
Each subject now stores how many tasks there are with this subject.
This works in the database implementations as follows:

- Setup a stream, on which the lists of subjects arrive
- Setup two streams, one listening to the subjects table and the other
one listening to the tasks table
- When the subjects stream updates, parse out the subjects list and add
it to the first stream
- When the tasks stream updates, query the subjects and add that list
to the first stream

This way, we basically listen to two streams at the same time, and
merge the results.