feat: Add class tests
This adds class tests to the app. Class tests are a second type of 
task holds a type ("Class Test" / "Vocabulary Test" / custom), a due date,
a reminder, a subject and a list of topics, each consisting of a topic and 
resources. In the resources field, the user can write down where they can
read up on the topic, where exercises are etc.
These class tests are merged into the list in the soon tasks widget on the 
left of the home screen, as well as the trash bin page, when they are deleted.

In addition, a class called AbstractTask has been created, which is the superclass
of ClassTest and Task. It defines functions both types of tasks need to have, and also
provides some default implementations.