feat: Add Task sharing
When the user now opens the shared task link obtained by clicking the
share icon in the ViewTaskPage, the app opens and tries loads the task.
If the the task's subject does not exist in the user's subjects, the
app will ask the user whether they want to create it. After that, they
are sent to the CreateTaskPage with all fields filled in, allowing the
user to potentially change some of the fields and finally create the
task.