feat(HomePage): Change layout on small screens
The HomePage layout now drastically changes on small screens. The new
layout only shows the SoonTasksWidget, however, it displays it's data
using a ListView instead of a GridView. This is supposed to make it
easier to read everything at a glance since there is no horizontal
scrolling involved anymore. The ListView shows ListTiles which show the
same information as the columns of the previously used grid.

The subjects have been moved into a page accessible through the
settings, which is always there no matter how big the screen is. This
page only shows the SubjectsList.

(This is sort of experimental. Definitely needs testing)